Only delete refresh token if access token present
In the frontend, the unauthenticate function should work regardless if there is
an access token available or not. Meaning, it should not crash, and correctly
clear the local storage and any token or account still kept, even if it can't
actually send the DELETE request to the API on the refresh token resource.

+++ frontend-service/src/authentication/UserContext.js
@@ -94,13 +94,13 @@ const UserProvider = ({ children }) => {
 
   const unauthenticate = async () => {
     if (refreshToken) {
-      const response = await (new AuthenticationService().deleteRefreshToken(refreshToken, accessToken));
-      if (response.status === 204) {
-        setRefreshToken(null);
-        localStorage.removeItem(refreshTokenKey);
-        setAccessToken(null);
-        setAccount(false);
+      if (accessToken) {
+        await (new AuthenticationService().deleteRefreshToken(refreshToken, accessToken));
       }
+      setRefreshToken(null);
+      localStorage.removeItem(refreshTokenKey);
+      setAccessToken(null);
+      setAccount(null);
     }
   };
